Splitted eval expression and flatter eval expression.
